move low_level::sprint_config to crate::config
